backend/requirements.txt
django>=5.0
djangorestframework
pandas
matplotlib
reportlab
django-cors-headers
requests
markdown
